requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
opencv-python-headless==4.11.0.86
pillow==11.1.0
pyzbar==0.1.9
qrcode==8.0
Werkzeug==3.1.3
